+++ mmm/mmm.h
@@ -20,7 +20,7 @@ void mmm_reset(double **);
 void mmm_freeup();
 void mmm_seq();
 void *mmm_par(void *args);
-double mmm_verify();
+float mmm_verify();
 void mmm_print();
 void mmm_print1(double **array);
